Rename BwTransaction to BwParser
